pyproject.toml
[project]
name = "tocadiscos"
version = "0.1.0"
description = "Music recommendation agent based on Last.fm history"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    # Core LLM & Agent Framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-mistralai>=0.2.0",
    "langchain-community>=0.3.0",
    # Observability
    "langfuse>=2.50.0",
    # HTTP & API
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    # Web search (for new releases discovery)
    "ddgs>=7.0.0",
    # Notifications
    "resend>=2.0.0",
    # Utils
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # Dev
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "ty>=0.0.1a6",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
python-version = "3.13"
